requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
async-generator==1.10
attrs==22.1.0
Automat==22.10.0
black==22.1.0
boto3==1.21.14
botocore==1.24.14
certifi==2022.9.24
cffi==1.15.1
cfscrape==2.1.1
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==38.0.1
cssselect==1.2.0
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
exceptiongroup==1.0.0
filelock==3.8.0
flake8==5.0.4
greenlet==1.1.3
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==0.10.0
lxml==4.9.1
mccabe==0.7.0
mypy-extensions==0.4.3
outcome==1.2.0
packaging==21.3
parsel==1.6.0
pathspec==0.9.0
Pillow==9.3.0
platformdirs==2.5.1
playwright==1.27.1
playwright-stealth==1.0.5
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.6
pyee==8.1.0
pyflakes==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
s3transfer==0.5.2
Scrapy==2.7.0
scrapy-cloudflare-middleware==0.0.1
scrapy-selenium==0.0.7
selenium==4.1.3
selenium-stealth==1.0.6
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tldextract==3.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.8
w3lib==2.0.1
websockets==10.4
whitenoise==6.0.0
wsproto==1.2.0
zope.interface==5.5.0
